Refactor OrcaHand into base and hardware layers
