+++ orca_core/__init__.py
@@ -5,5 +5,6 @@
 # You may use, copy, modify, and distribute this file under the terms of the MIT License.
 # See the LICENSE file at the root of this repository for full license information.
 # ==============================================================================
-from .core import OrcaHand
+from .core import BaseHand
 from .core import MockOrcaHand
+from .core import OrcaHand
