+++ orca_core/core.py
@@ -6,1027 +6,19 @@
 # See the LICENSE file at the root of this repository for full license information.
 # ==============================================================================
 
-import os
-import time
-import math
-import threading
-from typing import Dict, List, Union, TYPE_CHECKING
-from collections import deque
-from threading import RLock
-import numpy as np
-from .hardware.motor_client import MotorClient
-from .utils.utils import *
+from .base_hand import BaseHand
+from .hardware_hand import HardwareOrcaHand, MockOrcaHand
 
-if TYPE_CHECKING:
-    from .hardware.dynamixel_client import DynamixelClient
-    from .hardware.feetech_client import FeetechClient
 
-class OrcaHand:
-    """OrcaHand class is used to abtract hardware control the hand of the robot with simple high level control methods in joint space."""
-   
-    def __init__(self, model_path: str = None):
-        """Initialize the OrcaHand class.
+class OrcaHand(HardwareOrcaHand):
+    """Backward-compatible hardware-backed ORCA hand entry point."""
 
-        Args:
-            model_path (str): The path to model_path folder, which includes the config.yaml and calibration.yaml 
-        """
-        # Find the model directory if not provided
-        self.model_path = get_model_path(model_path)
-                
-        # Load configurations from the YAML files
-        self.config_path = os.path.join(self.model_path, "config.yaml")
-        self.calib_path = os.path.join(self.model_path, "calibration.yaml")
-        
-        config = read_yaml(self.config_path)
-        calib = read_yaml(self.calib_path)
-            
-        self.baudrate: int = config.get('baudrate', 3000000)
-        self.port: str = config.get('port', '/dev/ttyUSB0')
-        self.max_current: int = config.get('max_current', 300)
-        self.control_mode: str = config.get('control_mode', 'current_position')
-        self.type: str = config.get('type', None)
-        self.motor_type: str = config.get('motor_type', 'dynamixel')
-        
-        self.calib_current: str = config.get('calib_current', 200)
-        self.wrist_calib_current: str = config.get('wrist_calib_current', 100)
-        self.calib_step_size: float = config.get('calib_step_size', 0.1)
-        self.calib_step_period: float = config.get('calib_step_period', 0.01)
-        self.calib_threshold: float = config.get('calib_threshold', 0.01)
-        self.calib_num_stable: int = config.get('calib_num_stable', 20)
-        self.calib_sequence: Dict[str, Dict[str, str]] = config.get('calib_sequence', [])
-        self.calibrated: bool = calib.get('calibrated', False)
-        self.wrist_calibrated: bool = calib.get('wrist_calibrated', False)
-     
-        self.neutral_position: Dict[str, float] = config.get('neutral_position', {})
-        
-        self.motor_ids: List[int] = config.get('motor_ids', [])
-        self.joint_ids: List[str] = config.get('joint_ids', [])
-        self.motor_id_to_idx_dict: Dict[int, int] = {motor_id: i for i, motor_id in enumerate(self.motor_ids)}
 
-        motor_limits_from_calib_dict = calib.get('motor_limits', {})
-        self.motor_limits_dict: Dict[int, List[float]] = {
-            motor_id: motor_limits_from_calib_dict.get(motor_id, [None, None]) for motor_id in self.motor_ids}
-
-        joint_to_motor_ratios_from_calib_dict = calib.get('joint_to_motor_ratios', {})
-        self.joint_to_motor_ratios_dict: Dict[int, float] = {
-            motor_id: joint_to_motor_ratios_from_calib_dict.get(motor_id, 0.0) for motor_id in self.motor_ids}
-            
-        self.joint_to_motor_map: Dict[str, float] = config.get('joint_to_motor_map', {})
-        self.joint_roms_dict: Dict[str, List[float]] = config.get('joint_roms', {})
-        
-        self.joint_inversion_dict = {}
-        for joint, motor_id in self.joint_to_motor_map.items():
-            if motor_id < 0 or math.copysign(1, motor_id) < 0:
-                self.joint_inversion_dict[joint] = True
-                self.joint_to_motor_map[joint] = int(abs(motor_id))
-            else:
-                self.joint_inversion_dict[joint] = False
-        
-        self.joint_to_motor_map = {k: int(v) for k, v in self.joint_to_motor_map.items()} # This is to make IDs integers
-
-        self.motor_to_joint_dict: Dict[int, str] = {v: k for k, v in self.joint_to_motor_map.items()}
-
-        self._wrap_offsets_dict: Dict[int, float] = None
-
-        self._motor_client: MotorClient = None
-        self._motor_lock: RLock = RLock()
-
-        # Task thread to start and stop longer tasks like tensioning, calibration, etc. externally
-        self._task_thread: threading.Thread = None
-        self._task_stop_event = threading.Event()
-        self._lock = threading.Lock() 
-        self._current_task = None
-        
-        self._sanity_check()       
-        self.is_calibrated(verbose=True)
-
-    def __del__(self):
-        """Destructor to disconnect from the hand."""
-        self.disconnect()
-        
-    def _create_motor_client(self) -> MotorClient:
-        """Factory method to create the appropriate motor client based on motor_type config.
-
-        Returns:
-            MotorClient: The motor client instance.
-
-        Raises:
-            ValueError: If motor_type is not recognized.
-        """
-        if self.motor_type == 'dynamixel':
-            from .hardware.dynamixel_client import DynamixelClient
-            return DynamixelClient(self.motor_ids, self.port, self.baudrate)
-        elif self.motor_type == 'feetech':
-            from .hardware.feetech_client import FeetechClient
-            return FeetechClient(self.motor_ids, self.port, self.baudrate)
-        else:
-            raise ValueError(f"Unknown motor_type: {self.motor_type}. Expected 'dynamixel' or 'feetech'.")
-
-    def connect(self) -> tuple[bool, str]:
-        """Connect to the hand with the configured motor client.
-
-        Returns:
-            tuple[bool, str]: (Success status, message).
-        """
-        try:
-            self._motor_client = self._create_motor_client()
-            with self._motor_lock:
-                self._motor_client.connect()
-            return True, "Connection successful"
-        except Exception as e:
-            self._motor_client = None
-            print(f"Connection failed on {self.port}: {str(e)}")
-
-            # Try auto-detecting the port first
-            chosen_port = auto_detect_port()
-            if chosen_port and chosen_port != self.port:
-                try:
-                    self.port = chosen_port
-                    self._motor_client = self._create_motor_client()
-                    with self._motor_lock:
-                        self._motor_client.connect()
-                    update_yaml(self.config_path, 'port', chosen_port)
-                    return True, f"Connection successful with auto-detected port {chosen_port}"
-                except Exception:
-                    self._motor_client = None
-
-            # Fall back to interactive selection
-            print("Please select a port from available devices:")
-            chosen_port = get_and_choose_port()
-
-            if chosen_port is None:
-                return False, "Connection failed: No port selected"
-
-            # Try connecting with the chosen port
-            try:
-                self.port = chosen_port
-                self._motor_client = self._create_motor_client()
-                with self._motor_lock:
-                    self._motor_client.connect()
-
-                # Update config file with the working port
-                update_yaml(self.config_path, 'port', chosen_port)
-                return True, f"Connection successful with port {chosen_port}"
-            except Exception as e2:
-                self._motor_client = None
-                return False, f"Connection failed with selected port: {str(e2)}"
-        
-    def disconnect(self) -> tuple[bool, str]:
-        """Disconnect from the hand.
-
-        Returns:
-            tuple[bool, str]: (Success status, message).
-        """
-        try:
-            with self._motor_lock:
-                self.disable_torque()
-                time.sleep(0.1)
-                self._motor_client.disconnect()
-            return True, "Disconnected successfully"
-        except Exception as e:
-            return False, f"Disconnection failed: {str(e)}"
-        
-    def is_connected(self) -> bool:
-        """Check if the hand is connected.
-
-        Returns:
-            bool: True if connected, False otherwise.
-        """
-        return self._motor_client.is_connected if self._motor_client else False
-        
-    def enable_torque(self, motor_ids: List[int] = None):
-        """Enable torque for the motors.
-        
-        Args:
-            motor_ids (list): List of motor IDs to enable the torque. If None, all motors will be enabled
-        """
-        if motor_ids is None:
-            motor_ids = self.motor_ids
-        with self._motor_lock:
-            self._motor_client.set_torque_enabled(motor_ids, True)        
-
-    def disable_torque(self, motor_ids: List[int] = None):
-        """Disable torque for the motors.
-        
-        Args:
-            motor_ids (list): List of motor IDs to disable the torque. If None, all motors will be disabled.
-        """
-        if motor_ids is None:
-            motor_ids = self.motor_ids
-        with self._motor_lock:
-            self._motor_client.set_torque_enabled(motor_ids, False)
-    
-    def set_max_current(self, current: Union[float, List[float]]):
-        """Set the maximum current for the motors.
-        
-        Args:
-            current (int or list): If list, it should be the maximum current for each motor, otherwise it will be the same for all motors.
-        """
-        if isinstance(current, list):
-            if len(current) != len(self.motor_ids):
-                raise ValueError("Number of currents do not match the number of motors.")
-            with self._motor_lock:
-                self._motor_client.write_desired_current(self.motor_ids, current)
-        else:
-            with self._motor_lock:
-                self._motor_client.write_desired_current(self.motor_ids, current*np.ones(len(self.motor_ids)))
-        
-    def set_control_mode(self, mode: str, motor_ids: List[int] = None):
-        """Set the control mode for the motors.
-        
-        Args:
-            mode (str): Control mode.
-                (0) current: Current control mode,
-                (1) velocity: Velocity control mode,
-                (3) position: Position control mode,
-                (4) multi_turn_position: Multi-turn position control mode,
-                (5) current_based_position: Current-based position control mode.
-            motor_ids (list): List of motor IDs to set the control mode. If None, all motors will be set.
-            The wrist motor is controlled in multi-turn position control mode, if current or current-based position control mode is seleceted.
-        """
-        
-        mode_map = {
-            'current': 0,
-            'velocity': 1,
-            'position': 3,
-            'multi_turn_position': 4,
-            'current_based_position': 5
-        }
-
-        mode = mode_map.get(mode)
-        if mode is None:
-            raise ValueError("Invalid control mode.")
-        
-        with self._motor_lock:
-            if motor_ids is None:
-                motor_ids = self.motor_ids
-            else:
-                if not all(motor_id in self.motor_ids for motor_id in motor_ids):
-                    raise ValueError("Invalid motor IDs.")
-
-        if mode == 5 or mode == 0:
-            wrist_motor_id = self.joint_to_motor_map.get("wrist")
-            if wrist_motor_id is not None:
-                motor_ids_without_wrist = [motor_id for motor_id in motor_ids if motor_id != wrist_motor_id]
-                self._motor_client.set_operating_mode(motor_ids_without_wrist, mode)
-                if wrist_motor_id in motor_ids:
-                    self._motor_client.set_operating_mode([wrist_motor_id], 4)
-            else:
-                self._motor_client.set_operating_mode(motor_ids, mode)
-        else:
-            self._motor_client.set_operating_mode(motor_ids, mode)
-            
-    def get_motor_pos(self, as_dict: bool = False) -> Union[np.ndarray, dict]:
-        """Get the current motor positions in radians (Note that this includes offsets of the motors).
-        
-        Args:
-            as_dict (bool): If True, return the motor positions as a dictionary with motor IDs as keys.
-                           If False, return as numpy array.
-        
-        Returns:
-            Union[np.ndarray, dict]: Motor positions either as numpy array or dictionary {motor_id: position}.
-        """
-        with self._motor_lock:
-            motor_pos = self._motor_client.read_pos_vel_cur()[0]
-            if as_dict:
-                return {motor_id: pos for motor_id, pos in zip(self.motor_ids, motor_pos)}
-            return motor_pos
-        
-    def get_motor_current(self, as_dict: bool = False) -> Union[np.ndarray, dict]:
-        """Get the current motor currents in mA.
-        
-        Args:
-            as_dict (bool): If True, return the motor currents as a dictionary with motor IDs as keys.
-                           If False, return as numpy array.
-        
-        Returns:
-            Union[np.ndarray, dict]: Motor currents either as numpy array or dictionary {motor_id: current}.
-        """
-        with self._motor_lock:
-            motor_current = self._motor_client.read_pos_vel_cur()[2]
-            if as_dict:
-                return {motor_id: current for motor_id, current in zip(self.motor_ids, motor_current)}
-            return motor_current
-        
-    def get_motor_temp(self, as_dict: bool = False) -> Union[np.ndarray, dict]:
-        """Get the current motor temperatures in Celsius.
-        
-        Args:
-            as_dict (bool): If True, return the motor temperatures as a dictionary with motor IDs as keys.
-                           If False, return as numpy array.
-        
-        Returns:
-            Union[np.ndarray, dict]: Motor temperatures either as numpy array or dictionary {motor_id: temperature}.
-        """
-        with self._motor_lock:
-            motor_temp = self._motor_client.read_temperature()
-            if as_dict:
-                return {motor_id: temp for motor_id, temp in zip(self.motor_ids, motor_temp)}
-            return motor_temp
-
-    def get_joint_pos(self, as_list: bool = True) -> Union[dict, list]:
-        """Get the current joint positions.
-    
-        Args:
-            as_list (bool): If True, return the joint positions as a list in the order of joint_ids.
-                            If False, return the joint positions as a dictionary.
-    
-        Returns:
-            Union[dict, list]: Joint positions as a list [position1, position2, ...] in the order of joint_ids
-                               or as a dictionary {joint_name: position}.
-        """
-        motor_pos = self.get_motor_pos()
-        joint_pos = self._motor_to_joint_pos(motor_pos)
-        if as_list:
-            return [joint_pos[joint] for joint in self.joint_ids]
-    
-        return joint_pos
-         
-    def set_joint_pos(self, joint_pos: Union[dict, list], num_steps: int = 1, step_size: float = 1.0):
-        """Set the desired joint positions. If num_steps > 1, the hand will move to the target position in a smooth, gradual motion (depending also on step_size).
-    
-        Args:
-            joint_pos (dict or list): If dict, it should be {joint_name: desired_position}.
-                                    If list, it should contain positions in the order of joint_ids.
-            num_steps (int): Number of steps to reach the target position. If 1, moves directly to target.
-            step_size (float): Time to wait between steps in seconds.
-        """
-        
-        if num_steps > 1:
-            current_positions = self.get_joint_pos(as_list=False)
-            
-            if isinstance(joint_pos, list):
-                if len(joint_pos) != len(self.joint_ids):
-                    raise ValueError("Length of joint_pos list must match the number of joint_ids.")
-                target_positions = {joint: pos for joint, pos in zip(self.joint_ids, joint_pos)}
-            else:
-                target_positions = joint_pos.copy()
-            
-            for step in range(num_steps + 1):
-                t = step / num_steps
-                
-                interpolated_positions = {}
-                for joint in self.joint_ids:
-                    if joint in target_positions:
-                        current_pos = current_positions[joint]
-                        if current_pos is None or target_positions[joint] is None:
-                            interpolated_positions[joint] = None
-                            continue
-                        target_pos = target_positions[joint]
-                        interpolated_positions[joint] = current_pos * (1 - t) + target_pos * t
-                    else:
-                        interpolated_positions[joint] = current_positions[joint]
-                
-                motor_pos = self._joint_to_motor_pos(interpolated_positions)
-                                
-                self._set_motor_pos(motor_pos)
-                if step < num_steps: 
-                    time.sleep(step_size)
-        else:
-            if isinstance(joint_pos, dict):
-                motor_pos = self._joint_to_motor_pos(joint_pos)
-            elif isinstance(joint_pos, list):
-                if len(joint_pos) != len(self.joint_ids):
-                    raise ValueError("Length of joint_pos list must match the number of joint_ids.")
-                joint_pos_dict = {joint: pos for joint, pos in zip(self.joint_ids, joint_pos)}
-                motor_pos = self._joint_to_motor_pos(joint_pos_dict)
-            else:
-                raise ValueError("joint_pos must be a dict or a list.")
-
-            self._set_motor_pos(motor_pos)
-
-    def set_zero_position(self, num_steps: int = 25, step_size: float = 0.001):
-        """Set the hand to the zero position by moving all joints simultaneously to their zero positions
-        in a smooth, gradual motion.
-        
-        Args:
-            num_steps (int): Number of steps to reach the zero position.
-            step_size (float): Step size for each joint.
-        """
-        self.set_joint_pos({joint: 0 for joint in self.joint_ids}, num_steps=num_steps, step_size=step_size)
-        
-    def set_neutral_position(self, num_steps: int = 25, step_size: float = 0.001):
-        """Set the hand to the neutral position by moving all joints simultaneously to their neutral positions
-        in a smooth, gradual motion.
-
-        Args:
-            num_steps (int): Number of steps to reach the neutral position.
-            step_size (float): Step size for each joint.
-        """
-        if self.neutral_position is None:
-            raise ValueError("Neutral position is not set. Please set the neutral position in the config.yaml file.")
-        self.set_joint_pos(self.neutral_position, num_steps=num_steps, step_size=step_size)
-        
-    def init_joints(self, calibrate: bool = False
-                    ):
-        """Initialize the joints, enables torque, sets the control mode and sets to the zero position.
-        If the hand is not calibrated, it will calibrate the hand. 
-        
-        Args:
-            calibrate (bool): If True, the hand will be calibrated
-        
-        """
-        self.enable_torque()
-        self.set_control_mode(self.control_mode)
-        self.set_max_current(self.max_current)
-        
-        if not self.calibrated or calibrate:
-            self.calibrate()
-   
-        self._compute_wrap_offsets_dict()
-        self.set_joint_pos(self.neutral_position)
-
-    def is_calibrated(self, verbose: bool = False) -> bool:
-        """Check if the hand is calibrated.
-
-        Args:
-            verbose (bool): If True, print detailed calibration status for uncalibrated joints.
-
-        Returns:
-            bool: True if all joints are calibrated, False otherwise.
-        """
-        overall_calibrated = True
-        uncalibrated_messages = []
-        motors_with_warnings = set()
-
-        for motor_id, limits in self.motor_limits_dict.items():
-            if any(limit is None for limit in limits):
-                overall_calibrated = False
-                if not verbose:
-                    return False 
-                joint_name = self.motor_to_joint_dict.get(motor_id, "Unknown")
-                # Corrected escape sequence
-                uncalibrated_messages.append(
-                    f"\033[93mWarning: Motor ID {motor_id} (Joint: {joint_name}) has not been fully calibrated (missing motor limits).\033[0m")
-                motors_with_warnings.add(motor_id)
-
-
-        for motor_id, ratio in self.joint_to_motor_ratios_dict.items():
-            if ratio is None or ratio == 0.0:
-                overall_calibrated = False
-                if not verbose:
-                    return False
-                if motor_id not in motors_with_warnings:
-                    joint_name = self.motor_to_joint_dict.get(motor_id, "Unknown")
-                    uncalibrated_messages.append(
-                        f"\033[93mWarning: Motor ID {motor_id} (Joint: {joint_name}) has not been fully calibrated (missing joint-to-motor ratio).\033[0m"
-                    )
-                    motors_with_warnings.add(motor_id)
-        
-        if verbose:
-            for msg in uncalibrated_messages:
-                print(msg)
-        
-        return overall_calibrated
-
-    def calibrate(self, blocking: bool = True, force_wrist: bool = False):
-        if blocking:
-            self._task_stop_event.clear()
-            self._calibrate(force_wrist=force_wrist)
-        else:
-            self._start_task(self._calibrate, force_wrist=force_wrist)
-
-    def _calibrate(self, force_wrist: bool = False):
-
-        # Build effective calibration sequence with wrist logic
-        wrist_in_sequence = any('wrist' in step['joints'] for step in self.calib_sequence)
-        calib_sequence = list(self.calib_sequence)
-
-        if self.wrist_calibrated and not force_wrist:
-            if wrist_in_sequence:
-                print("WARNING: Wrist is already calibrated. Skipping wrist calibration. Use --force-wrist to override.")
-            calib_sequence = [step for step in calib_sequence if 'wrist' not in step['joints']]
-        elif not wrist_in_sequence:
-            calib_sequence.append({'step': len(calib_sequence) + 1, 'joints': {'wrist': 'flex'}})
-            calib_sequence.append({'step': len(calib_sequence) + 1, 'joints': {'wrist': 'extend'}})
-
-        # Store the min and max values for each motor
-        motor_limits = self.motor_limits_dict.copy()
-
-        self._compute_wrap_offsets_dict()
-        for step in calib_sequence:
-            for joint in step["joints"].keys():
-                motor_id = self.joint_to_motor_map[joint]
-                motor_limits[motor_id] = [None, None]
-                self._wrap_offsets_dict[motor_id] = 0.0
-
-        motors_with_initial_offset = set()
-        motors_with_final_offset = set()
-
-        # Set calibration control mode
-        self.set_control_mode('current_based_position')
-        self.set_max_current(self.calib_current)
-
-        for step in calib_sequence:
-
-            self.disable_torque()
-
-            if self._task_stop_event.is_set():
-                return
-
-            desired_increment, motor_reached_limit, directions, position_buffers, motor_reached_limit, calibrated_joints, position_logs, current_log = {}, {}, {}, {}, {}, {}, {}, {}
-
-            for joint, direction in step["joints"].items():
-
-                self.enable_torque(motor_ids=[self.joint_to_motor_map[joint]])
-
-                print("Enabling torque for the following motor: ", self.joint_to_motor_map[joint])
-
-                if self._task_stop_event.is_set():
-                    return
-
-                if joint == 'wrist':
-                    self.set_max_current(self.wrist_calib_current)
-                else:
-                    self.set_max_current(self.calib_current)
-
-                motor_id = self.joint_to_motor_map[joint]
-                sign = 1 if direction == 'flex' else -1
-                if self.joint_inversion_dict.get(joint, False):
-                    sign = -sign
-                directions[motor_id] = sign
-                position_buffers[motor_id] = deque(maxlen=self.calib_num_stable)
-                position_logs[motor_id] = []
-                current_log[motor_id] = []
-                motor_reached_limit[motor_id] = False
-
-                if self._motor_client.requires_offset_calibration and motor_id not in motors_with_initial_offset:
-                    self._motor_client.calibrate_offset(motor_id, upper=(sign < 0))
-                    motors_with_initial_offset.add(motor_id)
-
-
-
-            while(not all(motor_reached_limit.values()) and not self._task_stop_event.is_set()): 
-
-                for motor_id, reached_limit in motor_reached_limit.items():
-                    if not reached_limit:
-                        desired_increment[motor_id] = directions[motor_id] * self.calib_step_size
-
-                self._set_motor_pos(desired_increment, rel_to_current=True)
-                time.sleep(self.calib_step_period)
-                curr_pos = self.get_motor_pos()
-                
-                for motor_id in desired_increment.keys():
-                    if not motor_reached_limit[motor_id]:
-                        position_buffers[motor_id].append(curr_pos[self.motor_id_to_idx_dict[motor_id]])
-                        position_logs[motor_id].append(float(curr_pos[self.motor_id_to_idx_dict[motor_id]]))
-                        current_log[motor_id].append(float(self.get_motor_current()[self.motor_id_to_idx_dict[motor_id]]))
-
-                        # Check if limit found
-                        if len(position_buffers[motor_id]) == self.calib_num_stable and np.allclose(position_buffers[motor_id], position_buffers[motor_id][0], atol=self.calib_threshold):
-                            motor_reached_limit[motor_id] = True
-                            # disable torque for the motor
-                            if 'wrist' in joint:
-                                avg_limit = float(np.mean(position_buffers[motor_id]))
-                            else:
-                                self.disable_torque([motor_id])
-                                time.sleep(0.05)
-                                avg_limit = float(self.get_motor_pos()[self.motor_id_to_idx_dict[motor_id]])
-                            print(f"Motor {motor_id} corresponding to joint {self.motor_to_joint_dict[motor_id]} reached the limit at {avg_limit} rad.")
-                            if directions[motor_id] == 1:
-                                motor_limits[motor_id][1] = avg_limit
-                            if directions[motor_id] == -1:
-                                motor_limits[motor_id][0] = avg_limit
-
-                            # Set final offset after first limit found, then re-read limit in new coordinate system
-                            if self._motor_client.requires_offset_calibration and motor_id not in motors_with_final_offset:
-                                is_positive = directions[motor_id] > 0
-                                self._motor_client.calibrate_offset(motor_id, upper=is_positive)
-                                time.sleep(0.05)
-                                new_limit = float(self.get_motor_pos()[self.motor_id_to_idx_dict[motor_id]])
-                                motor_limits[motor_id][1 if is_positive else 0] = new_limit
-                                print(f"  (Offset adjusted: limit now at {new_limit} rad)")
-                                motors_with_final_offset.add(motor_id)
-
-                            self.enable_torque([motor_id])
-                
-            # find ratios of all motors that have been calibrated in this step
-            for joint, direction in step["joints"].items(): 
-                motor_id = self.joint_to_motor_map[joint]
-                if motor_limits[motor_id][0] is None or motor_limits[motor_id][1] is None:
-                    continue
-                delta_motor = motor_limits[motor_id][1] - motor_limits[motor_id][0]
-                delta_joint = self.joint_roms_dict[self.motor_to_joint_dict[motor_id]][1] - self.joint_roms_dict[self.motor_to_joint_dict[motor_id]][0]
-                self.joint_to_motor_ratios_dict[motor_id] = float(delta_motor / delta_joint) 
-                print("Joint calibrated: ", joint)
-                calibrated_joints[joint] = 0.0
-  
-            update_yaml(self.calib_path, 'joint_to_motor_ratios', self.joint_to_motor_ratios_dict)
-            update_yaml(self.calib_path, 'motor_limits', motor_limits)
-            self.motor_limits_dict = motor_limits
-            if calibrated_joints:
-                self.set_joint_pos(calibrated_joints, num_steps=25, step_size=0.001)
-            time.sleep(0.1)    
-            
-        # Update wrist_calibrated if wrist was calibrated in this run
-        if any('wrist' in step['joints'] for step in calib_sequence):
-            self.wrist_calibrated = True
-            update_yaml(self.calib_path, 'wrist_calibrated', True)
-
-        self.calibrated = self.is_calibrated()
-        update_yaml(self.calib_path, 'calibrated', self.calibrated)
-        self.set_joint_pos(calibrated_joints, num_steps=25, step_size=0.001)
-        self.set_max_current(self.max_current)
-       
-    def _compute_wrap_offsets_dict(self):
-        """Read motor_pos positions once and figure out ±1-turn offsets so that
-        pos + offset ∈ [motor_limits_lo, motor_limits_hi].
-        """
-
-        motor_pos = self.get_motor_pos()
-
-        lower_limit = np.array([self.motor_limits_dict[motor_id][0] for motor_id in self.motor_ids])
-        higher_limit = np.array([self.motor_limits_dict[motor_id][1] for motor_id in self.motor_ids])
-
-        offsets = {}
-        for i, motor_id in enumerate(self.motor_ids):
-            if lower_limit[i] is None or higher_limit[i] is None:
-                offsets[motor_id] = 0.0
-                continue
-
-            if motor_pos[i] < lower_limit[i] - 0.25 * np.pi: # Some buffer to compensate for noise/slack differences
-                print(f"Motor ID {motor_id} is out of bounds: "
-                    f"{lower_limit[i]} < {motor_pos[i]} < {higher_limit[i]}")
-                offsets[motor_id] = -2 * np.pi
-
-            elif motor_pos[i] > higher_limit[i] + 0.25 * np.pi: # Some buffer to compensate for noise/slack differences
-                print(f"Motor ID {motor_id} is out of bounds: "
-                    f"{lower_limit[i]} < {motor_pos[i]} < {higher_limit[i]}")
-                offsets[motor_id] = +2 * np.pi
-
-            else:
-                offsets[motor_id] = 0.0
-
-        print(f"Offsets: {offsets}")
-
-        self._wrap_offsets_dict = offsets
-
-    def _set_motor_pos(self, desired_pos: Union[dict, np.ndarray, list], rel_to_current: bool = False):
-        """Set the desired motor positions in radians.
-        
-        Args:
-            desired_pos (dict or np.ndarray or list): 
-                - If dict: {motor_id: desired_position}. Can be partial. Only motors in the dict will be commanded.
-                - If np.ndarray or list: Desired positions for all motors in the order of self.motor_ids.
-                                         None values will be skipped, and the corresponding motor won't be commanded.
-            rel_to_current (bool): If True, the desired position is relative to the current position.
-        """
-        with self._motor_lock:
-            if rel_to_current:
-                current_positions = self.get_motor_pos() # np.ndarray of all motor positions
-
-            motor_ids_to_write = []
-            positions_to_write = []
-
-            if isinstance(desired_pos, dict):
-                
-                for motor_id, pos_val in desired_pos.items():
-                    if motor_id not in self.motor_ids:
-                        print(f"Warning: Motor ID {motor_id} in desired_pos dict is not in self.motor_ids. Skipping.")
-                        continue
-
-                    if pos_val is None or math.isnan(pos_val):
-                        continue
-
-                    pos_to_write = float(pos_val)
-                    if rel_to_current:
-                        pos_to_write += current_positions[self.motor_id_to_idx_dict[motor_id]]
-                    
-                    motor_ids_to_write.append(motor_id)
-                    positions_to_write.append(pos_to_write)
-
-                if not motor_ids_to_write:
-                    return
-
-                positions_to_write = np.array(positions_to_write, dtype=float)
-
-            elif isinstance(desired_pos, (np.ndarray, list)):
-                if len(desired_pos) != len(self.motor_ids):
-                    raise ValueError(
-                        f"Length of desired_pos (list/ndarray) ({len(desired_pos)}) "
-                        f"must match the number of configured motor_ids ({len(self.motor_ids)})."
-                    )
-                
-                for i, pos_val in enumerate(desired_pos):
-                    if pos_val is None or math.isnan(pos_val):
-                        continue
-                    else:
-                        motor_ids_to_write.append(self.motor_ids[i])
-                        if rel_to_current:
-                            positions_to_write.append(float(pos_val) + current_positions[i])
-                        else:
-                            positions_to_write.append(float(pos_val))
-                
-                if not motor_ids_to_write:
-                    print(f"\033[93mWarning: All positions in desired_pos (list/array) were None. No motor commands sent.\033[0m")
-                    return
-
-                motor_ids_to_write = motor_ids_to_write
-                positions_to_write = np.array(positions_to_write, dtype=float)
-            
-            else:
-                raise ValueError("desired_pos must be a dict, np.ndarray, or list.")
-   
-            self._motor_client.write_desired_pos(motor_ids_to_write, positions_to_write)
-    
-    def _motor_to_joint_pos(self, motor_pos: np.ndarray) -> dict:
-        """Convert motor positions into joint positions.
-        
-        Args:
-            motor_pos (np.ndarray): Motor positions.
-        
-        Returns:
-            dict: {joint_name: position}
-        """
-        if self._wrap_offsets_dict is None:
-            self._compute_wrap_offsets_dict()
-
-        joint_pos = {}
-        for idx, pos in enumerate(motor_pos):
-            motor_id = self.motor_ids[idx]
-            joint_name = self.motor_to_joint_dict.get(motor_id)
-            if any(limit is None for limit in self.motor_limits_dict[motor_id]):
-                joint_pos[joint_name] = None #TODO: Add a warning here the probably the motor is not calibrated
-                print(f"\033[93mWarning: Motor ID {motor_id} (Joint: {joint_name}) has not been fully calibrated (missing motor limits).\033[0m")
-            elif self.joint_to_motor_ratios_dict[motor_id] == 0:
-                joint_pos[joint_name] = None #TODO: Add a warning here the probably the motor is not calibrated
-                print(f"\033[93mWarning: Motor ID {motor_id} (Joint: {joint_name}) has not been fully calibrated (missing joint-to-motor ratio).\033[0m")
-            else:
-                wrapped_pos = pos - self._wrap_offsets_dict.get(motor_id, 0.0)
-                
-                if self.joint_inversion_dict.get(joint_name, False):
-                    joint_pos[joint_name] = self.joint_roms_dict[joint_name][1] - (wrapped_pos - self.motor_limits_dict[motor_id][0]) / self.joint_to_motor_ratios_dict[motor_id]
-                else:
-                    joint_pos[joint_name] = self.joint_roms_dict[joint_name][0] + (wrapped_pos - self.motor_limits_dict[motor_id][0]) / self.joint_to_motor_ratios_dict[motor_id]
-        return joint_pos
-    
-    def _joint_to_motor_pos(self, joint_pos: dict) -> np.ndarray:
-        """Convert desired joint positions into motor commands.
-    
-        Args:
-            joint_pos (dict): {joint_name: desired_position}
-
-        Returns:
-            np.ndarray: Motor positions.
-        """
-        if self._wrap_offsets_dict is None:
-            self._compute_wrap_offsets_dict()
-
-        motor_pos = [None] * len(self.motor_ids)  # Initialize with None for all motors
-                
-        for joint_name, pos in joint_pos.items():
-            motor_id = self.joint_to_motor_map.get(joint_name)
-            if motor_id is None:
-                continue
-
-            if pos is None:
-                motor_pos[self.motor_id_to_idx_dict[motor_id]] = None
-                continue
-
-            if self.motor_limits_dict[motor_id][0] is None or self.motor_limits_dict[motor_id][1] is None or self.joint_to_motor_ratios_dict[motor_id] == 0:
-                motor_pos[self.motor_id_to_idx_dict[motor_id]] = None
-                print(f"\033[93mWarning: Motor ID {motor_id} (Joint: {joint_name}) has not been fully calibrated (missing joint-to-motor ratio).\033[0m")
-                continue
-            
-            min_pos, max_pos = self.joint_roms_dict[joint_name]
-            
-            if pos < min_pos or pos > max_pos:
-                clipped_pos = max(min_pos, min(max_pos, pos))
-                pos = clipped_pos
-
-            if self.joint_inversion_dict.get(joint_name, False):
-                # Inverted: higher ROM value corresponds to lower motor position.
-                motor_pos[self.motor_id_to_idx_dict[motor_id]] = self.motor_limits_dict[motor_id][0] + (self.joint_roms_dict[joint_name][1] - pos) * self.joint_to_motor_ratios_dict[motor_id]
-            else:
-                motor_pos[self.motor_id_to_idx_dict[motor_id]] = self.motor_limits_dict[motor_id][0] + (pos - self.joint_roms_dict[joint_name][0]) * self.joint_to_motor_ratios_dict[motor_id]  
-            
-            motor_pos[self.motor_id_to_idx_dict[motor_id]] += self._wrap_offsets_dict.get(motor_id, 0.0)
-            
-        return motor_pos
-    
-    def _sanity_check(self):
-        """Check if the configuration is correct and the IDs are consistent."""
-        if len(self.motor_ids) != len(self.joint_ids):
-            raise ValueError("Number of motor IDs and joints do not match.")
-        
-        if len(self.motor_ids) != len(self.joint_to_motor_map):
-            raise ValueError("Number of motor IDs and joints do not match.")
-        
-        if self.control_mode not in ['current_position', 'current_velocity', 'position', 'multi_turn_position', 'current_based_position']:
-            raise ValueError("Invalid control mode.")
-        
-        if self.max_current < self.calib_current:
-            raise ValueError("Max current should be greater than the calibration current.")
-                
-        for joint, motor_id in self.joint_to_motor_map.items():
-            if joint not in self.joint_ids:
-                raise ValueError(f"Joint {joint} is not defined.")
-            if joint not in self.joint_roms_dict:
-                raise ValueError(f"ROM for joint {joint} is not defined.")
-            if motor_id not in self.motor_ids:
-                raise ValueError(f"Motor ID {motor_id} is not in the motor IDs list.")
-            
-        for joint, rom in self.joint_roms_dict.items():
-            if rom[1] - rom[0] <= 0:
-                raise ValueError(f"ROM for joint {joint} is not valid.")
-            if joint not in self.joint_ids:
-                raise ValueError(f"Joint {joint} in ROMs is not defined.")
-            
-        for step in self.calib_sequence:
-            for joint, direction in step["joints"].items():
-                if joint not in self.joint_ids:
-                    raise ValueError(f"Joint {joint} is not defined.")
-                if direction not in ['flex', 'extend']:
-                    raise ValueError(f"Invalid direction for joint {joint}.")
-          
-        
-        for motor_limit in self.motor_limits_dict.values():
-            if any(limit is None for limit in motor_limit):
-                self.calibrated = False
-                update_yaml(self.calib_path, 'calibrated', False)
-
-    def tension(self, move_motors: bool = False, blocking: bool = True):
-        if blocking:
-            self._task_stop_event.clear()
-            self._tension(move_motors)
-        else:
-            self._start_task(self._tension, move_motors)
-
-    def jitter(self, motor_ids: List[int] = None, amplitude: float = 5.0,
-               frequency: float = 10.0, duration: float = 3.0,
-               include_wrist: bool = False, blocking: bool = True):
-        """Vibrate motors back and forth to release tension. Works at the motor level (no calibration required).
-
-        Args:
-            motor_ids (list): Motor IDs to jitter. If None, defaults to all motors (excluding wrist unless include_wrist is True).
-            amplitude (float): Vibration amplitude in degrees (max 10 for safety).
-            frequency (float): Vibration frequency in Hz.
-            duration (float): Duration of jitter in seconds.
-            include_wrist (bool): If True, include the wrist motor when motor_ids is None.
-            blocking (bool): If True, blocks until complete. If False, runs in a background thread.
-        """
-        if blocking:
-            self._task_stop_event.clear()
-            self._jitter(motor_ids, amplitude, frequency, duration, include_wrist)
-        else:
-            self._start_task(self._jitter, motor_ids, amplitude, frequency, duration, include_wrist)
-
-    def _jitter(self, motor_ids: List[int] = None, amplitude: float = 5.0,
-                frequency: float = 10.0, duration: float = 3.0,
-                include_wrist: bool = False):
-        MAX_AMPLITUDE_DEG = 10.0
-        if amplitude > MAX_AMPLITUDE_DEG:
-            raise ValueError(f"Amplitude must be <= {MAX_AMPLITUDE_DEG} degrees for safety. Got {amplitude}.")
-
-        amplitude_rad = np.deg2rad(amplitude)
-
-        if motor_ids is None:
-            wrist_motor_id = self.joint_to_motor_map.get("wrist")
-            motor_ids = [mid for mid in self.motor_ids if include_wrist or mid != wrist_motor_id]
-
-        start_positions = self.get_motor_pos(as_dict=True)
-        start_pos_array = np.array([start_positions[mid] for mid in motor_ids])
-
-        start_time = time.time()
-        while time.time() - start_time < duration and not self._task_stop_event.is_set():
-            t = time.time() - start_time
-            offset = amplitude_rad * math.sin(2 * math.pi * frequency * t)
-            with self._motor_lock:
-                self._motor_client.write_desired_pos(motor_ids, start_pos_array + offset)
-
-        with self._motor_lock:
-            self._motor_client.write_desired_pos(motor_ids, start_pos_array)
-
-    def _tension(self, move_motors: bool = False):
-        """Freeze the motors, so that the hand can be manually tensioned.
-        
-        Args:
-            move_motors (bool): If True, the hand will move to all motors positively for 3 seconds to set some initial tension.
-        """
-        self.set_control_mode('current_based_position')
-        if move_motors:
-            motors_to_move = [
-                motor_id for joint, motor_id in self.joint_to_motor_map.items()
-                if 'wrist' not in joint.lower() and motor_id in self.motor_ids
-            ]
-            self.set_max_current(self.calib_current)
-
-            duration = 8
-            increment_per_step = 0.1
-            motor_increments_right = {motor_id: increment_per_step for motor_id in motors_to_move}
-            motor_increments_left = {motor_id: -increment_per_step for motor_id in motors_to_move}
-
-            start_time = time.time()
-            while(time.time() - start_time < duration):
-                if self._task_stop_event.is_set():
-                    break
-                self._set_motor_pos(motor_increments_left, rel_to_current=True)
-                time.sleep(0.1)
-            
-            start_time = time.time()
-            while(time.time() - start_time < duration):
-                if self._task_stop_event.is_set():
-                    break
-                self._set_motor_pos(motor_increments_right, rel_to_current=True)
-                time.sleep(0.1)
-            
-
-
-        self.set_max_current(self.max_current)
-       
-        self.enable_torque()
-        print("Holding motors. Please tension carefully. Press Ctrl+C to exit.")
-        try:
-            while not self._task_stop_event.is_set():
-                time.sleep(0.1) 
-        finally:
-            self.disable_torque()  
-
-    def _run_task(self, task_fn, *args, **kwargs):
-        """Run a task in a separate thread, so that it can be stopped externally.
-        
-        Args:
-            task_fn (function): The task function to run.
-            *args: Additional arguments to pass to the task function.
-            **kwargs: Additional keyword arguments to pass to the task function.
-        """
-        with self._lock:
-            self._task_stop_event.clear()
-            self._current_task = task_fn.__name__
-            try:
-                task_fn(*args, **kwargs)
-            finally:
-                self._current_task = None
-
-    def _start_task(self, task_fn, *args, **kwargs):
-        """Start a task in a separate thread, so that it can be stopped externally.
-        
-        Args:
-            task_fn (function): The task function to run.
-            *args: Additional arguments to pass to the task function.
-            **kwargs: Additional keyword arguments to pass to the task function.
-        """
-        if self._task_thread and self._task_thread.is_alive():
-            print(f"Task '{self._current_task}' is already running.")
-            return
-
-        self._task_thread = threading.Thread(target=self._run_task, args=(task_fn,) + args, kwargs=kwargs)
-        self._task_thread.start()
-
-    def stop_task(self):
-        """Stop the currently running task.
-        """
-        if self._task_thread and self._task_thread.is_alive():
-            self._task_stop_event.set()
-            self._task_thread.join()
-            print("Task stopped.")
-        else:
-            print("No running task to stop.")               
-
-def require_connection(func):
-    def wrapper(self, *args, **kwargs):
-        if not self._motor_client.is_connected:
-            raise RuntimeError("Hand is not connected.")
-        return func(self, *args, **kwargs)
-    return wrapper
-
-def require_calibration(func):
-    def wrapper(self, *args, **kwargs):
-        if not self.calibrated:
-            raise RuntimeError("Hand is not calibrated. Please run .calibrate() first.")
-        return func(self, *args, **kwargs)
-    return wrapper
-
-
-class MockOrcaHand(OrcaHand):
-    """MockOrcaHand class is used to simulate the OrcaHand class for testing."""
-
-    def connect(self) -> tuple[bool, str]:
-        """Connects to the mock motor client.
-
-        Returns:
-            tuple[bool, str]: A tuple containing a boolean indicating success or failure,
-                              and a string message.
-        """
-        try:
-            from .hardware.mock_dynamixel_client import MockDynamixelClient
-            self._motor_client = MockDynamixelClient(self.motor_ids, self.port, self.baudrate)
-            with self._motor_lock:
-                self._motor_client.connect()
-            return True, "Mock connection successful"
-        except Exception as e:
-            self._motor_client = None
-            return False, f"Mock connection failed: {str(e)}"
-        
-    
 if __name__ == "__main__":
-    # Example usage:
     hand = OrcaHand()
     status = hand.connect()
     hand.enable_torque()
     hand.calibrate()
-
-    # Set the desired joint positions to 0
     hand.set_joint_pos({joint: 0 for joint in hand.joint_ids})
     hand.disable_torque()
     hand.disconnect()
